annotation_api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project]
name = "annotation-api"
version = "0.1.2.dev0"
description = "Backend for wildfire prevention, detection and monitoring"
authors = [
    { name = "Pyronear", email = "contact@pyronear.org" }
]
license = { text = "Apache-2.0" }
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.109.1,<1.0.0",
    "sqlmodel>=0.0.16",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "email-validator>=2.0.0",
    "requests>=2.32.0",
    "PyJWT>=2.8.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "uvicorn>=0.11.1,<1.0.0",
    "asyncpg>=0.25.0,<1.0.0",
    "alembic>=1.8.1",
    "sentry-sdk[fastapi]>=2.8.0",
    "posthog>=3.0.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "python-multipart==0.0.7",
    "python-magic>=0.4.17",
    "boto3>=1.26.0",
    "httpx>=0.24.0",
    "tqdm>=4.67.1",
    "rich>=13.0.0",
    "fastapi-pagination>=0.13.3",
]

[dependency-groups]
dev = [
    "mypy>=1.10.0",
    "ruff>=0.7.1",
    "pre-commit>=3.8.0",
    "types-requests>=2.0.0",
    "types-python-dateutil>=2.8.0",
    "sqlalchemy-stubs>=0.4",
    "types-passlib>=1.7.0",
    "pytest>=7.3.2,<9.0.0",
    "pytest-asyncio>=0.17.0,<1.0.0",
    "pytest-cov>=4.0.0",
    "pytest-pretty>=1.0.0",
    "requests-mock>=1.11.0",
    "httpx>=0.23.0",
    "aiosqlite>=0.16.0,<1.0.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
]

[tool.coverage.run]
source = ["src/app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
